dashboard/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
kafka-python==2.0.2
websockets==12.0